vamos-estudar/browser-tests/requirements.txt
browser-use==0.12.5
playwright==1.58.0
anthropic==0.76.0
python-dotenv==1.2.1
